+++ src/main/java/sample/database/dbo/RecipeDBO.java
@@ -178,7 +178,7 @@ public class RecipeDBO extends Model implements IRecipe {
     /**
      * {@inheritDoc}
      */
-    public void add(String ingredientName, int amount, String unitName) {
+    public void add(String ingredientName, double amount, String unitName) {
         final IngredientDAO ingredientDAO = new IngredientDAO();
         final UnitDAO unitDAO = new UnitDAO();
 
@@ -232,14 +232,14 @@ public class RecipeDBO extends Model implements IRecipe {
      * {@inheritDoc}
      */
     @Override
-    public List<Triple<IIngredient, Integer, IUnit>> getIngredients() {
-        List<Triple<IIngredient, Integer, IUnit>> ingredients = new ArrayList<>();
+    public List<Triple<IIngredient, Double, IUnit>> getIngredients() {
+        List<Triple<IIngredient, Double, IUnit>> ingredients = new ArrayList<>();
 
         final List<RecipeIngredientDBO> recipeIngredients = this.getAll(RecipeIngredientDBO.class);
 
         for (RecipeIngredientDBO recipeIngredient : recipeIngredients) {
             final IIngredient ingredient = recipeIngredient.parent(IngredientDBO.class);
-            final Integer amount = recipeIngredient.getInteger("amount");
+            final Double amount = recipeIngredient.getDouble("amount");
             final IUnit unit = recipeIngredient.parent(UnitDBO.class);
             ingredients.add(new ImmutableTriple<>(ingredient, amount, unit));
         }
@@ -254,7 +254,7 @@ public class RecipeDBO extends Model implements IRecipe {
      * @param recipeIngredient
      */
     @Override
-    public void add(Triple<IIngredient, Integer, IUnit> recipeIngredient) {
+    public void add(Triple<IIngredient, Double, IUnit> recipeIngredient) {
         this.add( recipeIngredient.getLeft().getName(), recipeIngredient.getMiddle(), recipeIngredient.getRight().getName() );
     }
